Have services show cluster name and link to service details page.
